+++ plugins/postcss-logical/CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changes to PostCSS Logical Properties
 
+### Unreleased (minor)
+
+- Added: Support for logical properties in `transition`
+
 ### 6.1.0 (February 15, 2023)
 
 - Adding support for `max-block-size`, `min-block-size`, `max-inline-size` and `min-inline-size`.